Add instructions to install AWS and gcloud CLI.

+++ install-gcloud.sh
@@ -1,32 +1,43 @@
 #!/bin/bash
 
-GPG_KEY_PATH=/usr/share/keyrings/gcloud.gpg
-CLUSTER_NAME=autopilot-cluster-1
-CLUSTER_REGION=us-central1
-PROJECT_ID=glossy-essence-352111
+# Check the latest version of gcloud CLI. https://cloud.google.com/sdk/docs/install-sdk
 
-# Install dependencies.
-sudo apt update
-sudo apt install -y \
-    apt-transport-https \
-    ca-certificates \
-    gnupg
+read -p "Which version of gcloud CLI would you like to install? " INPUT
+if [ -z $INPUT ]
+then
+    echo "Please enter a version number as the first argument."
+    exit 1
+else
+    # Set environment variables.
+    GCLOUD_VERSION=$INPUT
+    DOWNLOAD_FOLDER=~/downloads
+    INSTALL_FOLDER=/opt/google-cloud-sdk
+    CLUSTER_NAME=autopilot-cluster-1
+    CLUSTER_REGION=us-central1
+    PROJECT_ID=glossy-essence-352111
 
-# Download the repository key.
-curl https://packages.cloud.google.com/apt/doc/apt-key.gpg | sudo tee $GPG_KEY_PATH
+    # Clean up the directory of the same version.
+    sudo rm -r $INSTALL_FOLDER
+    sudo rm -r ~/.config/gcloud
+    
+    # Download the gcloud binary.
+    cd $DOWNLOAD_FOLDER
+    wget https://dl.google.com/dl/cloudsdk/channels/rapid/downloads/google-cloud-cli-$GCLOUD_VERSION-linux-x86_64.tar.gz
 
-# Create the repository configuration.
-echo "deb [signed-by=$GPG_KEY_PATH] https://packages.cloud.google.com/apt cloud-sdk main" | sudo tee /etc/apt/sources.list.d/gcloud.list > /dev/null
+    # Extract the binary file.
+    sudo mkdir $INSTALL_FOLDER
+    sudo tar -x -f google-cloud-cli-$GCLOUD_VERSION-linux-x86_64.tar.gz -C $INSTALL_FOLDER --strip-components=1
+    
+    # Install gcloud CLI.
+    cd $INSTALL_FOLDER
+    ./install.sh
 
-# Install the gcloud CLI from the repository.
-sudo apt update
-sudo apt install -y google-cloud-cli
+    # Initialize the gcloud CLI
+    gcloud init --no-browser
 
-# Initialize the gcloud CLI
-gcloud init --console-only
+    # Install kubectl through the gcloud CLI.
+    gcloud components install kubectl
 
-# Install kubectl through the gcloud CLI.
-gcloud components install kubectl
-
-# Configure kubectl command line access
-gcloud container clusters get-credentials $CLUSTER_NAME --region $CLUSTER_REGION --project $PROJECT_ID
+    # Configure kubectl command line access
+    gcloud container clusters get-credentials $CLUSTER_NAME --region $CLUSTER_REGION --project $PROJECT_ID
+fi
